# src/api/openapi.yaml
...
openapi: 3.0.0
info:
  title: Insights Remediations
  description: Insights Remediations Service
  version: 1.0.1
  contact:
    email: jozef@redhat.com
servers:
  - url: https://cloud.redhat.com/api/remediations/v1
    description: production server
  - url: /api/remediations/v1
    description: relative path
...
paths:
...
  /remediations/{id}/playbook_runs:
    get:
      tags:
        - remediations
      summary: List of executions of this remediation
      description: List of executions of this remediation
      operationId: listPlaybookRuns
      parameters:
        - $ref: '#/components/parameters/RemediationId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybookRunsList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - remediations
      summary: Execute remediation
      description: Execute remediation, RBAC permission {remediations:remediation:execute}
      operationId: runRemediation
      parameters:
        - $ref: '#/components/parameters/RemediationId'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecuteRemediation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
...
components:
  parameters:
    Limit:
      in: query
      name: limit
      description: Maximum number of results to return
      required: false
      schema:
        type: number
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      in: query
      name: offset
      description: Indicates the starting position of the query relative to the complete set of items that match the query
      required: false
      schema:
        type: number
        minimum: 0
        default: 0
...
    RemediationId:
      in: path
      name: id
      description: Remediation identifier
      required: true
      schema:
        $ref: '#/components/schemas/RemediationId'
...
  schemas:
...
    DateTime:
      type: string
      format: date-time
      example: '2018-12-05T08:19:36.641Z'
...
    RemediationId:
      type: string
      format: uuid
      example: 9197ba55-0abc-4028-9bbe-269e530f8bd5
...
    ExecutorId:
      type: string
      format: uuid
      example: 9197ba55-0abc-4028-9bbe-269e530f8bd5
    ExecutorName:
      type: string
      example: Executor Name
...
    PlaybookRunStatus:
      type: string
      example: "pending"
      enum: ["pending", "running", "success", "failure", "canceled"]
    PlaybookRunExecutorStatus:
      type: string
      example: "pending"
      enum: ["pending", "acked", "running", "success", "failure", "canceled"]
...
    PlaybookRunId:
      type: string
      format: uuid
      example: a8799a02-8be9-11e8-9eb6-529269fb1459
...
    RequestError:
      type: object
      additionalProperties: false
      required: [errors]
      properties:
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [id, status, code, title]
            properties:
              id:
                type: string
              status:
                type: integer
                example: 400
              code:
                type: string
                example: UNKNOWN_SYSTEM
              title:
                type: string
                example: Unknown system identifier 'abcd'
              details:
                type: object
    Meta:
      type: object
      additionalProperties: false
      required: [count, total]
      properties:
        count:
          type: number
          description: number of results returned
          example: 50
        total:
          type: number
          description: total number of results matching the query
          example: 114
    ExecuteRemediation:
      type: object
      additionalProperties: false
      required:
        - id
      properties:
        id:
          type: string
          example: a8799a02-8be9-11e8-9eb6-529269fb1459
    PlaybookRunsList:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PlaybookRunDetails'
        meta:
            $ref: '#/components/schemas/Meta'
    Counts:
      type: object
      additionalProperties: false
      required: [pending, running, success, failure, canceled]
      properties:
        pending:
          type: integer
        running:
          type: integer
        success:
          type: integer
        failure:
          type: integer
        canceled:
          type: integer
...
    ExecutorList:
      type: array
      items:
        $ref: '#/components/schemas/PlaybookExecutors'
...
    PlaybookRunDetails:
      type: object
      additionalProperties: false
      required:
        - id
        - executors
        - remediation_id
        - created_by
        - created_at
        - updated_at
        - status
      properties:
        id:
          $ref: '#/components/schemas/PlaybookRunId'
        executors:
          $ref: '#/components/schemas/ExecutorList'
        remediation_id:
          $ref: '#/components/schemas/RemediationId'
        created_by:
          $ref: '#/components/schemas/UserOut'
        created_at:
          $ref: '#/components/schemas/DateTime'
        updated_at:
          $ref: '#/components/schemas/DateTime'
        status:
          $ref: '#/components/schemas/PlaybookRunStatus'
...
    PlaybookExecutors:
      type: object
      additionalProperties: false
      required:
        - executor_id
        - executor_name
        - status
        - system_count
        - counts
      properties:
        executor_id:
          $ref: '#/components/schemas/ExecutorId'
        executor_name:
          $ref: '#/components/schemas/ExecutorName'
        status:
          $ref: '#/components/schemas/PlaybookRunExecutorStatus'
        system_count:
          type: integer
        counts:
          $ref: '#/components/schemas/Counts'
...
    UserOut:
      type: object
      additionalProperties: false
      required: [username, first_name, last_name]
      properties:
        username:
          type: string
          example: jharting@redhat.com
        first_name:
          type: string
          example: Jozef
        last_name:
          type: string
          example: Hartinger


    # Diagnosis specific types
...
  responses:
...
    NotFound:
      description: Entity Not Found
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RequestError'
...
    PreconditionFailed:
      description: One or more conditions given in the request header fields evaluated to false
...